+++ src/commands/adp/package/install.ts
@@ -15,9 +15,9 @@ import { SfdxCommand } from '@salesforce/command';
 import chalk from 'chalk';
 import childproc = require('child_process');
 import * as inquirer from 'inquirer';
+import { isNullOrUndefined } from 'util';
 import pkg = require('../../../helpers/packageHelper');
 import * as org from '../../../shared/orgUtil';
-import * as statusChecker from '../../../shared/statusChecker';
 
 export default class Install extends SfdxCommand {
     public static description = 'Installs the current package and/or its dependencies';
@@ -30,8 +30,10 @@ export default class Install extends SfdxCommand {
 
     protected static flagsConfig = {
         allpackages: { char: 'a', type: 'boolean', default: false, description: 'All packages, not just dependencies'},
-        forcelatest: {char: 'l', type: 'boolean', default: false,
-                        description: 'forces install of UN-released, latest package'},
+        versionbias: {
+            char: 'b', type: 'enum', description: 'Type of bias to use when determining package versions (Latest|Released)',
+            options: ['Latest', 'Released']
+        },
         noprompt: {type: 'boolean', description: 'disables all prompts'}
     };
 
@@ -62,18 +64,16 @@ export default class Install extends SfdxCommand {
         }
 
         const includeParent: boolean = this.flags.allpackages;
-        const forceLatest: boolean = this.flags.forcelatest;
 
         const verbose = true;
-        this.ux.startSpinner(chalk.gray('Retrieving package versions'));
+        console.log('Retrieving package versions');
         const projectJsonObj = await pkg.getProjectJson();
         const dependencies = await pkg.getDependencies({
             projectJson: projectJsonObj,
             includeParent,
             verbose,
-            latestVersions: forceLatest
+            versionBias: this.flags.versionbias
         });
-        this.ux.stopSpinner(chalk.gray('done.'));
 
         // TODO: Factor this out of the package commands and add to messages.
         if (dependencies === undefined) {
@@ -81,9 +81,9 @@ export default class Install extends SfdxCommand {
             return;
         }
 
-        this.ux.startSpinner(chalk.gray(`Retrieving installed dependencies from ${alias}`));
+        process.stdout.write(`Retrieving installed dependencies from ${alias}... `);
         const dependencyVersions = await pkg.getInstalledDependencies(dependencies, username);
-        this.ux.stopSpinner(chalk.gray('done.'));
+        console.log('done.');
 
         this.ux.log('Starting installation...');
 
@@ -99,27 +99,32 @@ export default class Install extends SfdxCommand {
                 this.ux.log(`Skipped ${pkgVersion.name} v${pkgVersion.installedVersion} ${releasedStr}. Already installed.`);
             } else {
                 // Install or upgrade the package
-                this.ux.log(`Installing ${pkgVersion.name} v${pkgVersion.latestVersion} ${releasedStr}...`);
-                const installCmd = `sfdx force:package:install -p ${pkgVersion.id} -u ${username} --json`;
-                const installRespJson = childproc.execSync(installCmd, {encoding: 'utf8'});
-                const installResp = JSON.parse(installRespJson);
-                const requestId: string = installResp.result.Id;
-                process.stdout.write('  Polling install status:    ');
-
-                const statusCmd = `sfdx force:package:install:report -i ${requestId}  -u ${username} --json`;
-
-                statusChecker.checkStatusUntilDone (statusCmd, () => { errors.push(pkgVersion.name); },
-                    { maxTries: 25, taskName: 'Installation'});
+                // Install or upgrade the package
+                const versionOrId = isNullOrUndefined(pkgVersion.installationVersion) ? pkgVersion.installationVersionId : `v${pkgVersion.installationVersion}`;
+                process.stdout.write(`Installing ${pkgVersion.name} ${versionOrId} ${releasedStr}... `);
+                const installCmd = `sfdx force:package:install -r -p ${pkgVersion.installationVersionId} -u ${username} --json --wait 20`;
+                let installRespJson;
+                try {
+                    installRespJson = childproc.execSync(installCmd, { encoding: 'utf8', stdio: 'pipe'});
+                    const installResp = JSON.parse(installRespJson);
+                    if (installResp.status === 0 && installResp.result.Status.toUpperCase() === 'SUCCESS') {
+                        console.log('done.');
+                    }
+                } catch (e) {
+                    console.log(chalk.red('failed.'));
+                    errors.push(installRespJson);
+                    break;
+                }
             }
         } // end for
 
         // Display end-of-process message
-        const endMsgStart: string = 'Installation completed';
+        const endMsgStart: string = 'Installation';
         let endMsg: string = '';
         if (errors.length > 0) {
-            endMsg = chalk.red(`${endMsgStart} with errors.`);
+            endMsg = chalk.red(`${endMsgStart} completed with errors.`);
         } else {
-            endMsg = chalk.green(`${endMsgStart} successfully.`);
+            endMsg = chalk.green(`${endMsgStart} completed successfully.`);
         }
         this.ux.log(endMsg);
     }
@@ -130,14 +135,18 @@ function buildConfirmMessage(flags, alias: string) {
     let message: string = '';
 
     if (!flags.noprompt) {
-        if (flags.allpackages && flags.forcelatest) {
-            message = `About to install ALL LATEST packages, including parent on ${chalk.magenta(alias)}. Confirm?`;
+        if (flags.allpackages && flags.versionbias === 'Latest') {
+            message = `About to install ALL LATEST packages, including parent on ${chalk.magenta(alias)}. Cotinue?`;
+        } else if (flags.allpackages && flags.versionbias === 'Released') {
+            message = `About to install ALL LATEST RELEASED package, including parent on ${chalk.magenta(alias)}. Continue?`;
+        } else if (flags.versionbias === 'Latest') {
+            message = `About to install LATEST DEPENDENCY packages on ${chalk.magenta(alias)}. Continue?`;
+        } else if (flags.versionbias === 'Released') {
+            message = `About to install LATEST RELEASED DEPENDENCY packages on ${chalk.magenta(alias)}. Continue?`;
         } else if (flags.allpackages) {
-            message = `About to install ALL RELEASED package, including parent on ${chalk.magenta(alias)}. Confirm?`;
-        } else if (flags.forcelatest) {
-            message = `About to install LATEST DEPENDENCY packages on ${chalk.magenta(alias)}. Confirm?`;
+            message = `About to install ALL packages AS CONFIGURED on ${chalk.magenta(alias)}. Continue?`;
         } else {
-            message = `About to install RELEASED DEPENDENCY packages on ${chalk.magenta(alias)}. Confirm?`;
+            message = `About to install DEPENDENCY packages AS CONFIGURED on ${chalk.magenta(alias)}. Continue?`;
         }
     }