+++ src/shared/cliCommand.ts
@@ -64,6 +64,11 @@ export function run(cmd: CLICommand, runner?) {
     }
 }
 
+/**
+ * Synchronously executes a command
+ * @param command the command (e.g. dx command) to execute
+ * @param noInherit boolean flag to omit stdio: 'inherit' option when calling execSync
+ */
 export function runCmd(command: string, noInherit?: boolean) {
     if (process.env.SYSTEM_DEBUG === 'true') {
         info(`Running CLI command: ${command}`);