+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Change Log
 0.0.4
 -----
   * Enhancements to package install command
+  * Replaced token parameter with username and password on tooling:query command
+  * Require SYSTEM_DEBUG=true to print CLI commands
 
 0.0.3
 -----